[sk] Reorganized file structure + fixed some nits
